+++ COLLADA2GLTF/shaders/commonProfileShaders.cpp
@@ -646,7 +646,7 @@ namespace GLTF
             } else {
                 shader->addUniform(symbol, type, count, forcesAsAnArray);
                 if ((count > 1) || forcesAsAnArray) {
-                    parameter->setUnsignedInt32(kCount, count);
+                    parameter->setUnsignedInt32(kCount, (int)count);
                 }
             }
